move alianzas to a module
