+++ gestion-de-alianzas/docker-compose.yml
@@ -12,9 +12,9 @@ services:
       - "5435:5432"
     healthcheck:
       test: ["CMD-SHELL", "pg_isready -U postgres_user -d gestion_alianzas -h 127.0.0.1"]
-      interval: 10s
+      interval: 5s
       timeout: 5s
-      retries: 5
+      retries: 10
       start_period: 30s
     volumes:
       - alianzas_pgdata:/var/lib/postgresql/data
