+++ gestion-de-alianzas/src/assembly.py
@@ -1,8 +1,8 @@
 # publicaciones_app/src/assembly.py
 
-from src.domain.use_cases.create_contrato_use_case import CreateContratoUseCase
-from src.domain.use_cases.base_use_case import BaseUseCase
-from src.adapters.postgres.contrato_postgres_adapter import PostgresContratoRepository
+from src.modulos.alianzas.domain.use_cases.create_contrato_use_case import CreateContratoUseCase
+from src.modulos.alianzas.domain.use_cases.base_use_case import BaseUseCase
+from src.modulos.alianzas.adapters.postgres.contrato_postgres_adapter import PostgresContratoRepository
 
 repository: PostgresContratoRepository = PostgresContratoRepository()
