+++ gestion-de-alianzas/src/modulos/alianzas/domain/models/contrato.py
