+++ gestion-de-alianzas/src/modulos/alianzas/domain/use_cases/base_use_case.py
