+++ gestion-de-alianzas/src/modulos/alianzas/infrastructure/db.py
